Gallery Images, Checkboxes, Thumbnail Width
Added background to all gallery posters (500x700) so that they are a uniform size. Used position to fix the alignment of the form's checkboxes with their respective type (It drove me crazy I hope that what I have is sufficient), Lastly, made the width of the thumbnail container be equal to the gallery main image.